Fix Double being represented as Long

+++ src/main/kotlin/net/model/Created.kt
@@ -1,12 +1,13 @@
 package net.model
 
+import java.math.BigDecimal
+
 interface Created {
     /**
      * The time of creation in local epoch-second format. ex: 1331042771.0
      *
      * Reddit says that the type is a [Long] but since it ends in ".0" you
-     * must store it as a double. However, in some tests, the response did not
-     * have a double, so maybe Long is fine? Todo: test if it's fine to use Long instead of Double.
+     * must store it as a [Double].
      */
-    val created: Long
+    val created: Double
 }
